refactor get_secret
