+++ litellm/batches/main.py
@@ -22,6 +22,7 @@ import litellm
 from litellm import client
 from litellm.llms.azure import AzureBatchesAPI
 from litellm.llms.openai import OpenAIBatchesAPI
+from litellm.secret_managers.main import get_secret
 from litellm.types.llms.openai import (
     Batch,
     CancelBatchRequest,
@@ -34,7 +35,7 @@ from litellm.types.llms.openai import (
     RetrieveBatchRequest,
 )
 from litellm.types.router import GenericLiteLLMParams
-from litellm.utils import get_secret, supports_httpx_timeout
+from litellm.utils import supports_httpx_timeout
 
 ####### ENVIRONMENT VARIABLES ###################
 openai_batches_instance = OpenAIBatchesAPI()
