+++ litellm/llms/bedrock/common_utils.py
@@ -10,7 +10,7 @@ from typing import List, Optional, Union
 import httpx
 
 import litellm
-from litellm import get_secret
+from litellm.secret_managers.main import get_secret
 
 
 class BedrockError(Exception):
