+++ litellm/proxy/proxy_config.yaml
@@ -1,15 +1,20 @@
 model_list:
- - model_name: fake-openai-endpoint
-   litellm_params:
-     model: openai/fake
-     api_key: fake-key
-   model_info:
-     id: "team-a-model" # used for identifying model in response headers
+  - model_name: fake-openai-endpoint
+    litellm_params:
+      model: openai/fake
+      api_base: https://exampleopenaiendpoint-production.up.railway.app/
+      api_key: os.environ/OPENAI_API_KEY
+  - model_name: gpt-3.5-turbo-end-user-test
+    litellm_params:
+      model: azure/chatgpt-v-2
+      api_base: https://openai-gpt-4-test-v-1.openai.azure.com/
+      api_version: "2023-05-15"
+      api_key: os.environ/AZURE_API_KEY
 
 litellm_settings:
   success_callback: ["prometheus"]
   failure_callback: ["prometheus"]
 
-general_settings: 
- master_key: sk-1234 
- custom_auth: example_config_yaml.custom_auth_basic.user_api_key_auth
+general_settings:
+  master_key: sk-1234
+  key_management_system: "google_secret_manager"
